// front/lib/generated/asset_registry/types/otcDealStatus.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum OtcDealStatus {
  Open,
  Completed,
  Expired,
  Cancelled,
}

export type OtcDealStatusArgs = OtcDealStatus;

export function getOtcDealStatusEncoder(): FixedSizeEncoder<OtcDealStatusArgs> {
  return getEnumEncoder(OtcDealStatus);
}

export function getOtcDealStatusDecoder(): FixedSizeDecoder<OtcDealStatus> {
  return getEnumDecoder(OtcDealStatus);
}

export function getOtcDealStatusCodec(): FixedSizeCodec<
  OtcDealStatusArgs,
  OtcDealStatus
> {
  return combineCodec(getOtcDealStatusEncoder(), getOtcDealStatusDecoder());
}