// front/lib/generated/asset_registry/types/vaultState.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum VaultState {
  Active,
  Triggered,
  Realized,
  Reverted,
  Expired,
  Returned,
}

export type VaultStateArgs = VaultState;

export function getVaultStateEncoder(): FixedSizeEncoder<VaultStateArgs> {
  return getEnumEncoder(VaultState);
}

export function getVaultStateDecoder(): FixedSizeDecoder<VaultState> {
  return getEnumDecoder(VaultState);
}

export function getVaultStateCodec(): FixedSizeCodec<
  VaultStateArgs,
  VaultState
> {
  return combineCodec(getVaultStateEncoder(), getVaultStateDecoder());
}