// front/lib/generated/asset_registry/types/vaultType.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum VaultType {
  Vesting,
  ConversionPending,
  DeliveryEscrow,
  RedemptionQueue,
}

export type VaultTypeArgs = VaultType;

export function getVaultTypeEncoder(): FixedSizeEncoder<VaultTypeArgs> {
  return getEnumEncoder(VaultType);
}

export function getVaultTypeDecoder(): FixedSizeDecoder<VaultType> {
  return getEnumDecoder(VaultType);
}

export function getVaultTypeCodec(): FixedSizeCodec<VaultTypeArgs, VaultType> {
  return combineCodec(getVaultTypeEncoder(), getVaultTypeDecoder());
}